+++ src/iop/tonecurve.c
@@ -781,8 +781,8 @@ static gboolean dt_iop_tonecurve_enter_notify(GtkWidget *widget, GdkEventCrossin
 {
   dt_iop_module_t *self = (dt_iop_module_t *)user_data;
   dt_iop_tonecurve_gui_data_t *c = (dt_iop_tonecurve_gui_data_t *)self->gui_data;
-  c->mouse_x = fabsf(c->mouse_x);
-  c->mouse_y = fabsf(c->mouse_y);
+  c->mouse_x = fabs(c->mouse_x);
+  c->mouse_y = fabs(c->mouse_y);
   gtk_widget_queue_draw(widget);
   return TRUE;
 }
@@ -792,8 +792,8 @@ static gboolean dt_iop_tonecurve_leave_notify(GtkWidget *widget, GdkEventCrossin
   dt_iop_module_t *self = (dt_iop_module_t *)user_data;
   dt_iop_tonecurve_gui_data_t *c = (dt_iop_tonecurve_gui_data_t *)self->gui_data;
   // weird sign dance for fluxbox:
-  c->mouse_x = -fabsf(c->mouse_x);
-  c->mouse_y = -fabsf(c->mouse_y);
+  c->mouse_x = -fabs(c->mouse_x);
+  c->mouse_y = -fabs(c->mouse_y);
   gtk_widget_queue_draw(widget);
   return TRUE;
 }
