+++ src/libs/live_view.c
@@ -621,7 +621,7 @@ gui_post_expose(dt_lib_module_t *self, cairo_t *cr, int32_t width, int32_t heigh
         x1=width;
         y1=y0;
       }
-      gboolean mouse_over_control = (lib->splitline_rotation%2==0)?(fabsf(sl_x-pointerx)<5):(fabsf(sl_y-pointery)<5);
+      gboolean mouse_over_control = (lib->splitline_rotation%2==0)?(fabs(sl_x-pointerx)<5):(fabs(sl_y-pointery)<5);
       cairo_save(cr);
       cairo_set_source_rgb(cr, .7, .7, .7);
       cairo_set_line_width(cr, (mouse_over_control ? 2.0 : 0.5) );
@@ -807,10 +807,10 @@ int button_pressed (struct dt_lib_module_t *self, double x, double y, double pre
     double sl_x = lib->overlay_x0 + lib->splitline_x * width;
     double sl_y = lib->overlay_y0 + lib->splitline_y * height;
 
-    gboolean mouse_over_control = (lib->splitline_rotation%2==0)?(fabsf(sl_x-x)<5):(fabsf(sl_y-y)<5);
+    gboolean mouse_over_control = (lib->splitline_rotation%2==0)?(fabs(sl_x-x)<5):(fabs(sl_y-y)<5);
 
     /* do the split rotating */
-    if(which==1 && fabsf(sl_x-x)<7 && fabsf(sl_y-y)<7)
+    if(which==1 && fabs(sl_x-x)<7 && fabs(sl_y-y)<7)
     {
       /* let's rotate */
       lib->splitline_rotation = (lib->splitline_rotation+1)%4;
